# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PAPELERIA LUNA",
        "description": "API de la Papelería con CRUD",
        "contact": {
            "email": "klebert_gabriel07@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Iniciar sesión",
                "description": "Inicia sesión para obtener un token Bearer que permita acceder a los recursos\n    protegidos.",
                "operationId": "7fe7085e1df995c6c2ee884a80eb30d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "correo",
                                    "contraseña",
                                    "dispositivo"
                                ],
                                "properties": {
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "klebert@admin.com"
                                    },
                                    "contraseña": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "dispositivo": {
                                        "type": "string",
                                        "example": "android"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token generado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/productos": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Consultar todos los productos",
                "description": "Retorna todos los Productos ",
                "operationId": "70a301af8ea04ce2f2596a58758632cd",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No se encontraron productos "
                    },
                    "405": {
                        "description": "Método no permitido"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Producto"
                ],
                "summary": "Crear Producto ",
                "description": "Crear un nuevo Producto ",
                "operationId": "4ff8851cb6bfe92aa226cdc804917ff0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "categoria",
                                    "precio"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Libreta del América"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Libreta de las águilas del América"
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "Cuadernos"
                                    },
                                    "precio": {
                                        "type": "integer",
                                        "example": "1000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto creado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/productos/{producto}": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Obtener un producto por su ID",
                "description": "Retorna un producto, con su Nombre, descripcion del producto, categoria a la que pertenece y precio ",
                "operationId": "e9473fa0579696e2bf18dbc66ded7556",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Producto"
                ],
                "summary": "Actualizar producto",
                "description": "Actualizar un producto por su ID",
                "operationId": "8dd1d809a11f298b920e852c19f33d98",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "descripcion",
                                    "categoria",
                                    "precio"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Libreta Águilas del América"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "libreta del bicampeón"
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "cuadernos"
                                    },
                                    "precio": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Producto actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Producto"
                ],
                "summary": "Eliminar Producto",
                "description": "Elimina un producto por su ID.",
                "operationId": "67c891f59ac40edfe3f3776b17497161",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "description": "ID del producto ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado con éxito"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": " no encontrada"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Login",
            "description": "Login"
        },
        {
            "name": "Producto",
            "description": "Producto"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum Bearer Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}